refactoring openfoam export
