+++ porespy/io/__init__.py
@@ -13,12 +13,14 @@ io
     porespy.io.to_palabos
     porespy.io.to_openpnm
     porespy.io.openpnm_to_im
+    porespy.io.im_to_openfoam
 
 .. autofunction:: to_vtk
 .. autofunction:: dict_to_vtk
 .. autofunction:: to_palabos
 .. autofunction:: to_openpnm
 .. autofunction:: openpnm_to_im
+.. autofunction:: im_to_openfoam
 
 
 """
@@ -28,3 +30,4 @@ from .__funcs__ import dict_to_vtk
 from .__funcs__ import to_palabos
 from .__funcs__ import to_openpnm
 from .__funcs__ import openpnm_to_im
+from .__funcs__ import im_to_openfoam
